requirements.txt
neo4j
requests
python-dotenv
pandas
numpy
plotly
kaleido
python-arango